requirements.txt
requests~=2.32.4
python-dotenv~=1.1.0
fastapi[standard]~=0.115.13
pydantic~=2.11.7
pandas